Experiments with refcounted type
The packages in refcount/ are meant as a replacement to GNATCOLL.Refcount.
The latter has a few drawbacks: it requires the type to be tagged and
derived from Refcounted; it might be doing too many allocation when we
want to store simple types like Integer for instance.
Reusing the traits approach might less us better control and optimize
memory allocation.
Setting the value of a million smart pointer is indeed faster when using
traits for a simple type like Integer; it is however slower when the
object is unconstrained, presumably because there is one more call to
malloc().
The Adjust operation seems to be consistently faster, although it isn't
clear why (unless we are in the case were we do not use atomic operations).
Getting the value from the smart pointer is much slower than in GNATCOLL,
and this isn't clear why either (5 times slower). Perhaps hardware cache
is better in the case of the GNATCOLL approach.