Split the large conts-lists package into multiple small packages
The goal is to remove the useless with clauses when possible (most
notably the with on Ada.Finalization which might not be correct in a
qualification context).

Pass a storage_pool argument to multiple generic packages, to let user
control how the memory allocation is performed (paving the way for having
more efficient allocations, or perhaps persistent containers via a
storage pool using mmap to allocate its buffer).

Remove the systematic use of Reference_Type is element traits. Instead,
we use a Return_Type, which in some cases will be an actual reference
type, and in other cases will be a copy of the element.

Use library projects to build (and separate test framework in its own
directory). This ensures that all files build correctly, even if not used
by the test framework.