Tests are now automatically generated.
They contain a lot of duplication, and we want to be able to test all
sorts of combination for the policies, so generating the tests automatically
eases that process.
The output is no longer done on stdout, since that has to be too concise.
Instead, we output the full information in a JSON file.