(cursors): rename from Element_Type to Return_Type
This is more consistent with the rest of the framework, and helps
clarify that we are not dealing with the customer's original type,
but some intermediate type defined by the elements traits packages.

Work in progress to get algorithms to work fine when a Return_Type
is defined as a Reference_Type (for efficiency).