Initial support for algorithms usable with reference_type
An algorithm expects a predicate that takes a Return_Type as a
parameter (since this is the data available from a cursor). But
this is not convenient in practice. For instance, when we store
strings in a container, we expect to be able to pass a predicate
that receives a String, not some reference_type for a string (and
the type is declared in some elements traits packages that most
users will have a hard time to find in any case).

The current approach is to have a function to convert from
Return_Type to Element_Type, that is passed to the algorithm (rather
than defined in the elements traits, since the algorithm does not
have access to the elements traits). But this extra function is then
returning an unconstrained type (a String) which is inefficient. As
such, the count_if test is taking 285% of the time it takes for C++
for instance.