Graphs: use cursors rather than For_Each_* subprograms
This is more efficient, since we avoid subprogram calls, even if it is
a bit more work to setup in some cases. It is also compatible with
the cursors used for the other containers, and therefore standard
algorithms like Count_If can also be used on graphs.

J212-008 and P112-064