Merge C++ graph tests with the Ada graph tests
So that the HTML page contains performance tests for both languages.

Display timing for graph tests in HTML page, as for other containers
Add support, in the HTML, to define specific tests for each category
of container, rather than have a hard-coded list of tests. Tests for
graphs are really different from the others.

Fix resizing issue with vectors: even after calling Reserve_Capacity,
we were reseting the size to 4 the first time an element was inserted,
thus making the call to Reserve_Capacity useless.
We now use the same policy as in C++: a vector is never shrunk when
elements are removed from it, except on an explicit call to
Shrink_To_Fit.
This improves the performance. In particular, a non-recursive DFS
implementation on a graph is now just as far as a recursive
implementation, when the temporary vector's capacity is set to the
number of elemenents in the graph, since we no longer have calls
to realloc (we still have one call to alloc initially).