Remove Incidence_Graph_Traits, merged with Conts.Graphs.Traits
This removes one level of instantiation, so is slightly easier to
use.

Also remove one level of instantiation for property maps, by using
the "others => <>" notation in formal package parameters

Move property maps in their own top-level package, since they can
be used for any type of container. We no longer distinguish between
exterior and interior maps in the packages, since they have the
same profiles. The algorithms however still need to know which is
which, so that they can create exterior maps as appropriate.