Instantiate cursor packages on parent type
When we provide predefined instantiations of standard containers
(like a definite-unbounded vector for instance), we have to define a
child type so that we can attach aspects information to support the
for-of loop.
However, we now instantiate the Cursors package on the type defined
in vectors-generics, rather than on each of our specific child type.
This removes the need for redeclaring some of the operations
locally (like Swap for instance) and algorithms like Shuffle are
thus easier to instantiate as shown in test_shuffle.adb