Separate cursors and getters
Algorithms now require two parameters: one for the cursor (forward,
bidirectional or random_access), and a getter (mapping the cursor to
some element type).
This means that in most cases, a single cursor traits is needed for
each container, since traversing is always done the same.
On the other hand, this lets users control how to retrieve the elements
(as the original Element_Type, or as the Returned_Type which might be
more efficient but sometimes harder to use). In particular, this removes
the needs for the Count_If_Convert package, and all similar variants that
would have been needed for other algorithms.

An extra flexibility is added by using property maps instead of a simpler
getter function, since now a separate type is responsible for retrieving
the data from the container. Although in most cases it will be the
container itself (and appropriate simpler versions of the algorithm are
used), this opens up possibilities, like retrieving information from
multiple sources to pass it to one algorithm.