Move the Cursors, Element_Maps and returned_Maps packages in *.Generics
This makes a lot of packages easier, in particular for users who
instantiate their own lists, since they no longer to define the
Iterable aspect or cursor traits separately.

This required adding an "Impl" package to *.Generics, since instantiating
the cursor traits need a full view of the container and cursor types. On
the other hand, this ensures that only subprograms that are meant to be
used in application code are left in the public API (and not for
instance the First_Primitive, Next_Primitive,...)