Return a reference_type for indefinite types, whenever possible
These are much more efficient that returning a copy of the element,
but wasn't possible to do before in all cases because the
Constant_Indexing aspect must return a reference type with an
access-to-constant discriminant.

Remove the conts-elements-indefinite_ref package, since it was now
the same as conts-elements-indefinite. Remove associated
predefined containers