Maps.Resize: now manipulates Count_Type
Resize used to deal with Hash_Type, but this is the wrong type
semantically.

Change the definition of Conts.Count_Type to an explicit range type,
rather than the platform and implementation-specific Natural. This is
so that a Count_Type always fits in a Hash_Type.
This however required a lot of cleaning up in packages that were
assuming we had a straight Integer. This is a nice cleanup, since
it means we could eventually redefine those types easily now.

This explicit type also means that user code cannot assume any more
that Conts.Count_Type is the same as Ada.Containers.Count_Type, which
again is good if we ever decide to redefine those types.

Change the use of Integer in use_* packages to use an new Element_Type
instead, for clarity.

Change-Id: I3e372851c720173ec973d4975fe85609e81e35f2