Make Last work for empty vectors
For this, we need to have an Extended_Index, as done in the standard
Ada containers, which has one more value to indicate invalid position
in the vector. That way, one can still write
   for Idx in Index_Type'First .. Vec.Last loop
even when the vector is empty.

A similar restriction applies to graphs adjacency lists, since they
are based on vectors.
The example was updated to show how this can be used with enumerations

Change-Id: I04fc85a4fb773cb835ccc025c24821a80153c32a