Split bounded lists and vectors into SPARK and Ada implementations
The SPARK implementation cannot be made controlled, since it isn't
possible to specify the base container type (we can't add a
discriminant in a tagged extension). So it really is only compatible
with definite elements which use the default assignment behavior.
The SPARK version only works for definite elements, so the formal
parameter is not directly the traits package itself.

But for Ada, we would like to support any type of element traits
(indefinite or user-defined), which means we need to have a
controlled list, and thus we need to specify the base container
type.

For now, the simplest is to have two implementations of the
bounded storage.

The high-level package Definite_Bounded lists is based on the SPARK
implementation

Other approaches were tried, for instance using a controlled
component added later in the list hierarchy, but that forces the
list to be limited.

Change-Id: Ib3e62c403779ef8ebb5e681c94f73de80c23aaf1